==> 6_Figure6_BiofilmASVSharing.Rmd <==
---
title: "Figure 6 - Biofilm ASV shairng"
author: "Sven Tobias-Hunefeldt"
date: "03/04/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
#Set up
Import data and refine as done using the set up script

```{r load packages}

library(ggpubr)
library(ggplot2)
library(Rmisc)
library(vegan)
library(tidyr)
library(dplyr)
library(zetadiv)
library(glm2)
library(tidyverse)
library(rstatix)
library(broom)


```
```{r Refine colours and shapes}
#Refine colours
cbbPalette <- c("black",
                "red",
                "navy", 
                "green", 
                "magenta",
                "forestgreen",
                "turquoise")

Mesh_Palette = c("Enclosed" = "#000000",
                 "Non-enclosed" = "#E69F00")

Time_colour_list = c("0" = "turquoise", 
                     "7" = "forestgreen", 
                     "14" = "black", 
                     "19" = "red", 
                     "28" = "navy", 
                     "42" = "green", 
                     "56" = "magenta")

#Refine shape list
Substrate_shape_list = c("Plastic" = 15, #Filled square 
                         "Glass" = 16, #Filled circle
                         "Tile" = 17, #Filled Triangle
                         "Wood" = 5, #Hollow diamond
                         "Water" = 11) #Lined star
#Make colour lists
Substrate_colour_list = c("Plastic" = "black",
                          "Glass" = "green",
                          "Tile" = "blue",
                          "Wood" = "magenta")

```
#Figure 6 A
##Set up and run ASV
```{r}
####Prokaryotes ####
#Extract otu table and convert to presence absence scheme
otutable_df = as.data.frame(as.matrix(otu_table(Phyloseq_Biofilm_16S_noA)))
otutable_df[otutable_df>0]<-1

#Make empty dataframes for each time point
otutable_df_D7_enc = data.frame()
otutable_df_D14_enc = data.frame()
otutable_df_D19_enc = data.frame()
otutable_df_D28_enc = data.frame()
otutable_df_D42_enc = data.frame()
otutable_df_D56_enc = data.frame()
otutable_df_D7_expo = data.frame()
otutable_df_D14_expo = data.frame()
otutable_df_D19_expo = data.frame()
otutable_df_D28_expo = data.frame()
otutable_df_D42_expo = data.frame()
otutable_df_D56_expo = data.frame()

#Fill dataframes
i = 1
#Split original dataframe into separate dataframes based on sample days
for (i in 1:nrow(otutable_df)) {
if (grepl("1", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){ #Loop for 1 in sample name, and if true run associated code, it not true then move onto next else.
  otutable_df_D7_enc = rbind(otutable_df_D7_enc, otutable_df[i,]) #Extract row and combine into day specific dataframe.
  
} else if (grepl("2", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){
  otutable_df_D14_enc = rbind(otutable_df_D14_enc, otutable_df[i,])
  
} else if (grepl("3", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){
  otutable_df_D19_enc = rbind(otutable_df_D19_enc, otutable_df[i,])
  
} else if (grepl("4", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){
  otutable_df_D28_enc = rbind(otutable_df_D28_enc, otutable_df[i,])
  
} else if (grepl("5", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){
  otutable_df_D42_enc = rbind(otutable_df_D42_enc, otutable_df[i,])
  
} else if (grepl("6", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){
  otutable_df_D56_enc = rbind(otutable_df_D56_enc, otutable_df[i,])

} else if (grepl("1", rownames(otutable_df)[i]) == T && grepl("e", rownames(otutable_df)[i]) == T){
  otutable_df_D7_expo = rbind(otutable_df_D7_expo, otutable_df[i,])
  
} else if (grepl("2", rownames(otutable_df)[i]) == T && grepl("e", rownames(otutable_df)[i]) == T){
  otutable_df_D14_expo = rbind(otutable_df_D14_expo, otutable_df[i,])
  
} else if (grepl("3", rownames(otutable_df)[i]) == T && grepl("e", rownames(otutable_df)[i]) == T){
  otutable_df_D19_expo = rbind(otutable_df_D19_expo, otutable_df[i,])
  
} else if (grepl("4", rownames(otutable_df)[i]) == T && grepl("e", rownames(otutable_df)[i]) == T){
  otutable_df_D28_expo = rbind(otutable_df_D28_expo, otutable_df[i,])
  
} else if (grepl("5", rownames(otutable_df)[i]) == T && grepl("e", rownames(otutable_df)[i]) == T){
  otutable_df_D42_expo = rbind(otutable_df_D42_expo, otutable_df[i,])
  
} else if (grepl("6", rownames(otutable_df)[i]) == T && grepl("e", rownames(otutable_df)[i]) == T){
  otutable_df_D56_expo = rbind(otutable_df_D56_expo, otutable_df[i,])
  
} 
  i=i+1
}

#Check rownames to see if it worked properly
rownames(otutable_df)
rownames(otutable_df_D7_enc)
rownames(otutable_df_D14_enc)
rownames(otutable_df_D19_enc)
rownames(otutable_df_D28_enc)
rownames(otutable_df_D42_enc)
rownames(otutable_df_D56_enc)
rownames(otutable_df_D7_expo)
rownames(otutable_df_D14_expo)
rownames(otutable_df_D19_expo)
rownames(otutable_df_D28_expo)
rownames(otutable_df_D42_expo)
rownames(otutable_df_D56_expo)



#Specify order number with however many samples there are
#Make lists
mesh_list = c("enc", "expo")
Time_list = c(7,
             14,
             19,
             28,
             42,
             56)
countordermax = list()
general_count = 1

#Count number of samples
for (i in 1:6) {
  for (y in 1:2){
    Object = get(paste0("otutable_df_D", Time_list[i], "_", mesh_list[y]))
    
    temp = nrow(Object)
    countordermax[general_count] = temp
    general_count = general_count + 1
  }
 }

countordermax

general_count = 1
  
i=1
y=1

#Set up progress bar and run calculations
pb = txtProgressBar(min = 0, max = length(countordermax), initial = 0, style = 3)
TIME <- Sys.time()
TIME
for (i in 1:6) {
  for (y in 1:2){
    Object = get(paste0("otutable_df_D", Time_list[i], "_", mesh_list[y]))
    zeta_decline.mc_calc = Zeta.decline.mc(data.spec =  Object, order = 1:countordermax[[general_count]])
    saveRDS(zeta_decline.mc_calc, paste0("otutable_df_16S_D", Time_list[i], "_", mesh_list[y]))

    Object2 = zeta_decline.mc_calc
      assign(paste0("otutable_df_D", Time_list[i], "_", mesh_list[y], "_16S_out"), Object2)
      
      setTxtProgressBar(pb, general_count)
      Sys.time() - TIME
      general_count = general_count + 1 
  }
  
}
Sys.time() - TIME

####Eukaryotes ####
#Extract otu table and convert to presence absence scheme
otutable_df = as.data.frame(as.matrix(otu_table(Phyloseq_Biofilm_18S_noA)))
otutable_df[otutable_df>0]<-1

#Make empty dataframes for each time point
otutable_df_D7_enc = data.frame()
otutable_df_D14_enc = data.frame()
otutable_df_D19_enc = data.frame()
otutable_df_D28_enc = data.frame()
otutable_df_D42_enc = data.frame()
otutable_df_D56_enc = data.frame()
otutable_df_D7_expo = data.frame()
otutable_df_D14_expo = data.frame()
otutable_df_D19_expo = data.frame()
otutable_df_D28_expo = data.frame()
otutable_df_D42_expo = data.frame()
otutable_df_D56_expo = data.frame()

#Fill dataframes
i = 1
#Split original dataframe into separate dataframes based on sample days
for (i in 1:nrow(otutable_df)) {
if (grepl("1", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){ #Loop for 1 in sample name, and if true run associated code, it not true then move onto next else.
  otutable_df_D7_enc = rbind(otutable_df_D7_enc, otutable_df[i,]) #Extract row and combine into day specific dataframe.
  
} else if (grepl("2", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){
  otutable_df_D14_enc = rbind(otutable_df_D14_enc, otutable_df[i,])
  
} else if (grepl("3", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){
  otutable_df_D19_enc = rbind(otutable_df_D19_enc, otutable_df[i,])
  
} else if (grepl("4", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){
  otutable_df_D28_enc = rbind(otutable_df_D28_enc, otutable_df[i,])
  
} else if (grepl("5", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){
  otutable_df_D42_enc = rbind(otutable_df_D42_enc, otutable_df[i,])
  
} else if (grepl("6", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){
  otutable_df_D56_enc = rbind(otutable_df_D56_enc, otutable_df[i,])

} else if (grepl("1", rownames(otutable_df)[i]) == T && grepl("e", rownames(otutable_df)[i]) == T){
  otutable_df_D7_expo = rbind(otutable_df_D7_expo, otutable_df[i,])
  
} else if (grepl("2", rownames(otutable_df)[i]) == T && grepl("e", rownames(otutable_df)[i]) == T){
  otutable_df_D14_expo = rbind(otutable_df_D14_expo, otutable_df[i,])
  
} else if (grepl("3", rownames(otutable_df)[i]) == T && grepl("e", rownames(otutable_df)[i]) == T){
  otutable_df_D19_expo = rbind(otutable_df_D19_expo, otutable_df[i,])
  
} else if (grepl("4", rownames(otutable_df)[i]) == T && grepl("e", rownames(otutable_df)[i]) == T){
  otutable_df_D28_expo = rbind(otutable_df_D28_expo, otutable_df[i,])
  
} else if (grepl("5", rownames(otutable_df)[i]) == T && grepl("e", rownames(otutable_df)[i]) == T){
  otutable_df_D42_expo = rbind(otutable_df_D42_expo, otutable_df[i,])
  
} else if (grepl("6", rownames(otutable_df)[i]) == T && grepl("e", rownames(otutable_df)[i]) == T){
  otutable_df_D56_expo = rbind(otutable_df_D56_expo, otutable_df[i,])
  
} 
  i=i+1
}

#Check rownames to see if it worked properly
#rownames(otutable_df)
rownames(otutable_df_D7_enc)
rownames(otutable_df_D14_enc)
rownames(otutable_df_D19_enc)
rownames(otutable_df_D28_enc)
rownames(otutable_df_D42_enc)
rownames(otutable_df_D56_enc)
rownames(otutable_df_D7_expo)
rownames(otutable_df_D14_expo)
rownames(otutable_df_D19_expo)
rownames(otutable_df_D28_expo)
rownames(otutable_df_D42_expo)
rownames(otutable_df_D56_expo)


#Specify order number with however many samples there are
mesh_list = c("enc", "expo")
Time_list = c(7,
             14,
             19,
             28,
             42,
             56)
countordermax = list()
general_count = 1
 
#Count number of samples
for (i in 1:6) {
  for (y in 1:2){
    Object = get(paste0("otutable_df_D", Time_list[i], "_", mesh_list[y]))
    
    temp = nrow(Object)
    countordermax[general_count] = temp
    general_count = general_count + 1
  }
 }

countordermax

general_count = 1
  
i=1
y=1

#Set up progress bar and run calculations
pb = txtProgressBar(min = 0, max = length(countordermax), initial = 0, style = 3)
TIME <- Sys.time()
TIME
for (i in 1:6) {
  for (y in 1:2){
    Object = get(paste0("otutable_df_D", Time_list[i], "_", mesh_list[y]))
    zeta_decline.mc_calc = Zeta.decline.mc(data.spec =  Object, order = 1:countordermax[[general_count]])
    saveRDS(zeta_decline.mc_calc, paste0("otutable_df_18S_D", Time_list[i], "_", mesh_list[y]))

    Object2 = zeta_decline.mc_calc
      assign(paste0("otutable_df_D", Time_list[i], "_", mesh_list[y], "_18S_out"), Object2)
      
      setTxtProgressBar(pb, general_count)
      Sys.time() - TIME
      general_count = general_count + 1 
  }
  
}
Sys.time() - TIME
```
##Extract data
```{r}
#Make extraction lists
Time_list = c(7,14,19,28,42,56)
mesh_list = c("enc", "expo")
Org_list = c("16S", "18S")

#Carry out extraction
Summary_Zeta = data.frame(matrix(ncol = 5, nrow = 0))
x = c("Day", "Mesh_status", "Shared", "Unique", "Sequencing")
colnames(Summary_Zeta) <- x
#pb = txtProgressBar(min = 0, max = length(Time_list), initial = 0)
i = 1
for (i in 1:length(Time_list)) {
  for (y in 1:length(mesh_list)) {
    for (x in 1:length(Org_list)) {
    
    assign("temp", get(paste0("otutable_df_D", Time_list[i], "_", mesh_list[y], "_", Org_list[x], "_out")))
    
    temp_df = data.frame("Day" = Time_list[i], 
                         "Mesh_status" = mesh_list[y],
                         "Shared" = tail(temp$zeta.val, n = 1), 
                         "Unique" = head(temp$zeta.val, n = 1), 
                         "Sequencing" = Org_list[x])
    
    Summary_Zeta = rbind(Summary_Zeta, temp_df)

#setTxtProgressBar(pb,i)
    }
  }
}
Summary_Zeta


Summary_Zeta$Mesh_status = gsub("enc", "Enclosed", Summary_Zeta$Mesh_status)
Summary_Zeta$Mesh_status = gsub("expo", "Non-enclosed", Summary_Zeta$Mesh_status)

Summary_Zeta
```
##Plot
```{r}
#Convert to factor and reorder for easier plotting, so plotting makes logical sense, and consistency
Summary_Zeta$Day = as.factor(as.character(Summary_Zeta$Day))
Summary_Zeta$Day = factor(Summary_Zeta$Day,
                         levels=c("7",
                                  "14",
                                  "19",
                                  "28",
                                  "42",
                                  "56"))
Summary_Zeta$Sequencing = as.character(Summary_Zeta$Sequencing)
Summary_Zeta$Sequencing = gsub("16S", "Prokaryotes", Summary_Zeta$Sequencing)
Summary_Zeta$Sequencing = gsub("18S", "Eukaryotes", Summary_Zeta$Sequencing)
Summary_Zeta$Sequencing = factor(Summary_Zeta$Sequencing,
                         levels=c("Prokaryotes", 
                                  "Eukaryotes"))

#Make plot
Shared_plot_substrate = ggplot(data = Summary_Zeta, aes(x = Day, y = Shared, colour = Mesh_status, group = Mesh_status))+
  geom_line(lwd = 1)+
  scale_color_manual("Condition", values = Mesh_Palette)+
  scale_y_log10()+
  xlab("Time (days)")+
  ylab("Number of shared ASVs")+
 # ggtitle("Shared ASVs across samples increases with time")+
  facet_grid(. ~ Sequencing)+
  theme_bw()+
  My_Theme

#Show plot
Shared_plot_substrate

```

##Calculate mean trends and statistical significance
```{r Set up}

#Subset to prokaryotes and eukaryotes specifically
Summary_Zeta_Prok = subset(Summary_Zeta, Sequencing == "Prokaryotes")
Summary_Zeta_Euk = subset(Summary_Zeta, Sequencing == "Eukaryotes")

#Add early vs late as new column
Summary_Zeta_Prok$StageofBiofilm = "tmp"
Summary_Zeta_Prok$StageofBiofilm[grep("7", Summary_Zeta_Prok$Day)] = "Early"
Summary_Zeta_Prok$StageofBiofilm[grep("14", Summary_Zeta_Prok$Day)] = "Early"
Summary_Zeta_Prok$StageofBiofilm[grep("19", Summary_Zeta_Prok$Day)] = "Late"
Summary_Zeta_Prok$StageofBiofilm[grep("28", Summary_Zeta_Prok$Day)] = "Late"
Summary_Zeta_Prok$StageofBiofilm[grep("42", Summary_Zeta_Prok$Day)] = "Late"
Summary_Zeta_Prok$StageofBiofilm[grep("56", Summary_Zeta_Prok$Day)] = "Late"
Summary_Zeta_Euk$StageofBiofilm = "tmp"
Summary_Zeta_Euk$StageofBiofilm[grep("7", Summary_Zeta_Euk$Day)] = "Early"
Summary_Zeta_Euk$StageofBiofilm[grep("14", Summary_Zeta_Euk$Day)] = "Late"
Summary_Zeta_Euk$StageofBiofilm[grep("19", Summary_Zeta_Euk$Day)] = "Late"
Summary_Zeta_Euk$StageofBiofilm[grep("28", Summary_Zeta_Euk$Day)] = "Late"
Summary_Zeta_Euk$StageofBiofilm[grep("42", Summary_Zeta_Euk$Day)] = "Late"
Summary_Zeta_Euk$StageofBiofilm[grep("56", Summary_Zeta_Euk$Day)] = "Late"

Summary_Zeta = rbind(Summary_Zeta_Prok, Summary_Zeta_Euk)
```
```{r Calculate means}
test_mean = Summary_Zeta_Prok %>% 
  group_by(Day) %>% 
  summarise(mean_val = mean(Shared), sd = sd(Shared))
print(test_mean)

#Factor increase from 7 to 56
test_mean$mean_val[6]/test_mean$mean_val[1]

test_mean = Summary_Zeta_Euk %>% 
  group_by(Day) %>% 
  summarise(mean_val = mean(Shared), sd = sd(Shared))
print(test_mean)

#Factor increase from 7 to 56
test_mean$mean_val[6]/test_mean$mean_val[1]
```
```{r Stats}
kruskal.test(Summary_Zeta$Shared, Summary_Zeta$Day, p.adjust.method = "bonferroni")
#Non-significant
wilcox.test(Summary_Zeta$Shared ~ Summary_Zeta$StageofBiofilm, p.adjust.method = "bonferroni")
#Significant

kruskal.test(Summary_Zeta_Prok$Shared, Summary_Zeta_Prok$Day, p.adjust.method = "bonferroni")
#Non-significant
kruskal.test(Summary_Zeta_Euk$Shared, Summary_Zeta_Euk$Day, p.adjust.method = "bonferroni")
#Non-significant

wilcox.test(Summary_Zeta_Prok$Shared ~ Summary_Zeta_Prok$StageofBiofilm, p.adjust.method = "bonferroni")
#Significant
wilcox.test(Summary_Zeta_Euk$Shared ~ Summary_Zeta_Euk$StageofBiofilm, p.adjust.method = "bonferroni")
#Non-significant

pairwise.wilcox.test(Summary_Zeta_Prok$Shared, Summary_Zeta_Prok$Day, p.adjust.method = "bonferroni")
#No significance
pairwise.wilcox.test(Summary_Zeta_Euk$Shared, Summary_Zeta_Euk$Day, p.adjust.method = "bonferroni")
#No significance

cor.test(as.numeric(as.character(Summary_Zeta_Prok$Shared)), as.numeric(as.character(Summary_Zeta_Prok$Day)), method = "s")
#Significant
cor.test(as.numeric(as.character(Summary_Zeta_Euk$Shared)), as.numeric(as.character(Summary_Zeta_Euk$Day)), method = "s")
#Non-significant
```
#Figure 6 B
##Set up and run ASV
```{r}

####Prokaryotes ####
#Extarct otu table and convert to presence/absence scheme
otutable_df = as.data.frame(as.matrix(otu_table(Phyloseq_Biofilm_16S_noA)))
otutable_df[otutable_df>0]<-1


#Make empty dataframes for each time point
otutable_df_D7_plastic = data.frame()
otutable_df_D14_plastic = data.frame()
otutable_df_D19_plastic = data.frame()
otutable_df_D28_plastic = data.frame()
otutable_df_D42_plastic = data.frame()
otutable_df_D56_plastic = data.frame()
otutable_df_D7_glass = data.frame()
otutable_df_D14_glass = data.frame()
otutable_df_D19_glass = data.frame()
otutable_df_D28_glass = data.frame()
otutable_df_D42_glass = data.frame()
otutable_df_D56_glass = data.frame()
otutable_df_D7_tile = data.frame()
otutable_df_D14_tile = data.frame()
otutable_df_D19_tile = data.frame()
otutable_df_D28_tile = data.frame()
otutable_df_D42_tile = data.frame()
otutable_df_D56_tile = data.frame()
otutable_df_D7_wood = data.frame()
otutable_df_D14_wood = data.frame()
otutable_df_D19_wood = data.frame()
otutable_df_D28_wood = data.frame()
otutable_df_D42_wood = data.frame()
otutable_df_D56_wood = data.frame()

#Fill dataframes
i = 1
#Split original dataframe into separate dataframes based on sample days
for (i in 1:nrow(otutable_df)) {
if (grepl("1a", rownames(otutable_df)[i]) == T){ #Loop for 1 in sample name, and if true run associated code, it not true then move onto next else.
  otutable_df_D7_plastic = rbind(otutable_df_D7_plastic, otutable_df[i,]) #Extract row and combine into day specific dataframe.
  
} else if (grepl("2a", rownames(otutable_df)[i]) == T){
  otutable_df_D14_plastic = rbind(otutable_df_D14_plastic, otutable_df[i,])
  
} else if (grepl("3a", rownames(otutable_df)[i]) == T){
  otutable_df_D19_plastic = rbind(otutable_df_D19_plastic, otutable_df[i,])
  
} else if (grepl("4a", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){
  otutable_df_D28_plastic = rbind(otutable_df_D28_plastic, otutable_df[i,])
  
} else if (grepl("5a", rownames(otutable_df)[i]) == T){
  otutable_df_D42_plastic = rbind(otutable_df_D42_plastic, otutable_df[i,])
  
} else if (grepl("6a", rownames(otutable_df)[i]) == T){
  otutable_df_D56_plastic = rbind(otutable_df_D56_plastic, otutable_df[i,])

} else if (grepl("1g", rownames(otutable_df)[i]) == T){
  otutable_df_D7_glass = rbind(otutable_df_D7_glass, otutable_df[i,]) 
  
} else if (grepl("2g", rownames(otutable_df)[i]) == T){
  otutable_df_D14_glass = rbind(otutable_df_D14_glass, otutable_df[i,])
  
} else if (grepl("3g", rownames(otutable_df)[i]) == T){
  otutable_df_D19_glass = rbind(otutable_df_D19_glass, otutable_df[i,])
  
} else if (grepl("4g", rownames(otutable_df)[i]) == T){
  otutable_df_D28_glass = rbind(otutable_df_D28_glass, otutable_df[i,])
  
} else if (grepl("5g", rownames(otutable_df)[i]) == T){
  otutable_df_D42_glass = rbind(otutable_df_D42_glass, otutable_df[i,])
  
} else if (grepl("6g", rownames(otutable_df)[i]) == T){
  otutable_df_D56_glass = rbind(otutable_df_D56_glass, otutable_df[i,])

} else if (grepl("1t", rownames(otutable_df)[i]) == T){
  otutable_df_D7_tile = rbind(otutable_df_D7_tile, otutable_df[i,]) 
  
} else if (grepl("2t", rownames(otutable_df)[i]) == T){
  otutable_df_D14_tile = rbind(otutable_df_D14_tile, otutable_df[i,])
  
} else if (grepl("3t", rownames(otutable_df)[i]) == T){
  otutable_df_D19_tile = rbind(otutable_df_D19_tile, otutable_df[i,])
  
} else if (grepl("4t", rownames(otutable_df)[i]) == T){
  otutable_df_D28_tile = rbind(otutable_df_D28_tile, otutable_df[i,])
  
} else if (grepl("5t", rownames(otutable_df)[i]) == T){
  otutable_df_D42_tile = rbind(otutable_df_D42_tile, otutable_df[i,])
  
} else if (grepl("6t", rownames(otutable_df)[i]) == T){
  otutable_df_D56_tile = rbind(otutable_df_D56_tile, otutable_df[i,])

} else if (grepl("1p", rownames(otutable_df)[i]) == T){
  otutable_df_D7_wood = rbind(otutable_df_D7_wood, otutable_df[i,]) 
  
} else if (grepl("2p", rownames(otutable_df)[i]) == T){
  otutable_df_D14_wood = rbind(otutable_df_D14_wood, otutable_df[i,])
  
} else if (grepl("3p", rownames(otutable_df)[i]) == T){
  otutable_df_D19_wood = rbind(otutable_df_D19_wood, otutable_df[i,])
  
} else if (grepl("4p", rownames(otutable_df)[i]) == T){
  otutable_df_D28_wood = rbind(otutable_df_D28_wood, otutable_df[i,])
  
} else if (grepl("5p", rownames(otutable_df)[i]) == T){
  otutable_df_D42_wood = rbind(otutable_df_D42_wood, otutable_df[i,])
  
} else if (grepl("6p", rownames(otutable_df)[i]) == T){
  otutable_df_D56_wood = rbind(otutable_df_D56_wood, otutable_df[i,])

} 
  i=i+1
}

#Check rownames to see if it worked properly
rownames(otutable_df)
rownames(otutable_df_D7_plastic)
rownames(otutable_df_D14_plastic)
rownames(otutable_df_D19_plastic)
rownames(otutable_df_D28_plastic)
rownames(otutable_df_D42_plastic)
rownames(otutable_df_D56_plastic)
rownames(otutable_df_D7_glass)
rownames(otutable_df_D14_glass)
rownames(otutable_df_D19_glass)
rownames(otutable_df_D28_glass)
rownames(otutable_df_D42_glass)
rownames(otutable_df_D56_glass)
rownames(otutable_df_D7_tile)
rownames(otutable_df_D14_tile)
rownames(otutable_df_D19_tile)
rownames(otutable_df_D28_tile)
rownames(otutable_df_D42_tile)
rownames(otutable_df_D56_tile)
rownames(otutable_df_D7_wood)
rownames(otutable_df_D14_wood)
rownames(otutable_df_D19_wood)
rownames(otutable_df_D28_wood)
rownames(otutable_df_D42_wood)
rownames(otutable_df_D56_wood)



#Specify order number with however many samples there are
substrate_list = c("plastic",
              "glass",
              "tile",
              "wood")
Time_list = c(7,
             14,
             19,
             28,
             42,
             56)
countordermax = list()
general_count = 1
#Run calculations
for (y in 1:4){
  for (i in 1:6) {
    Object = get(paste0("otutable_df_D", Time_list[i], "_", substrate_list[y]))
    
    temp = nrow(Object)
    countordermax[general_count] = temp
    general_count = general_count + 1
  }
 }

countordermax

general_count = 1
  
i=1
y=1

#Make progress bar and run ASV sharing calculations
pb = txtProgressBar(min = 0, max = length(countordermax), initial = 0, style = 3)
TIME <- Sys.time()
TIME
for (y in 1:4){
  for (i in 1:6) {
    Object = get(paste0("otutable_df_D", Time_list[i], "_", substrate_list[y]))
    zeta_decline.mc_calc = Zeta.decline.mc(data.spec =  Object, order = 1:countordermax[[general_count]])
    saveRDS(zeta_decline.mc_calc, paste0("otutable_df_16S_D", Time_list[i], "_", substrate_list[y]))

    Object2 = zeta_decline.mc_calc
      assign(paste0("otutable_df_D", Time_list[i], "_", substrate_list[y], "_16S_out"), Object2)
      
      setTxtProgressBar(pb, general_count)
      Sys.time() - TIME
      general_count = general_count + 1 
  }
  
}
Sys.time() - TIME

####Eukaryotes ####
#Extarct otu table and convert to presence/absence scheme
otutable_df = as.data.frame(as.matrix(otu_table(Phyloseq_Biofilm_18S_noA)))
otutable_df[otutable_df>0]<-1


#Make empty dataframes for each time point
otutable_df_D7_plastic = data.frame()
otutable_df_D14_plastic = data.frame()
otutable_df_D19_plastic = data.frame()
otutable_df_D28_plastic = data.frame()
otutable_df_D42_plastic = data.frame()
otutable_df_D56_plastic = data.frame()
otutable_df_D7_glass = data.frame()
otutable_df_D14_glass = data.frame()
otutable_df_D19_glass = data.frame()
otutable_df_D28_glass = data.frame()
otutable_df_D42_glass = data.frame()
otutable_df_D56_glass = data.frame()
otutable_df_D7_tile = data.frame()
otutable_df_D14_tile = data.frame()
otutable_df_D19_tile = data.frame()
otutable_df_D28_tile = data.frame()
otutable_df_D42_tile = data.frame()
otutable_df_D56_tile = data.frame()
otutable_df_D7_wood = data.frame()
otutable_df_D14_wood = data.frame()
otutable_df_D19_wood = data.frame()
otutable_df_D28_wood = data.frame()
otutable_df_D42_wood = data.frame()
otutable_df_D56_wood = data.frame()

#Fill dataframes
i = 1
#Split original dataframe into separate dataframes based on sample days
for (i in 1:nrow(otutable_df)) {
if (grepl("1a", rownames(otutable_df)[i]) == T){ #Loop for 1 in sample name, and if true run associated code, it not true then move onto next else.
  otutable_df_D7_plastic = rbind(otutable_df_D7_plastic, otutable_df[i,]) #Extract row and combine into day specific dataframe.
  
} else if (grepl("2a", rownames(otutable_df)[i]) == T){
  otutable_df_D14_plastic = rbind(otutable_df_D14_plastic, otutable_df[i,])
  
} else if (grepl("3a", rownames(otutable_df)[i]) == T){
  otutable_df_D19_plastic = rbind(otutable_df_D19_plastic, otutable_df[i,])
  
} else if (grepl("4a", rownames(otutable_df)[i]) == T && grepl("m", rownames(otutable_df)[i]) == T){
  otutable_df_D28_plastic = rbind(otutable_df_D28_plastic, otutable_df[i,])
  
} else if (grepl("5a", rownames(otutable_df)[i]) == T){
  otutable_df_D42_plastic = rbind(otutable_df_D42_plastic, otutable_df[i,])
  
} else if (grepl("6a", rownames(otutable_df)[i]) == T){
  otutable_df_D56_plastic = rbind(otutable_df_D56_plastic, otutable_df[i,])

} else if (grepl("1g", rownames(otutable_df)[i]) == T){
  otutable_df_D7_glass = rbind(otutable_df_D7_glass, otutable_df[i,]) 
  
} else if (grepl("2g", rownames(otutable_df)[i]) == T){
  otutable_df_D14_glass = rbind(otutable_df_D14_glass, otutable_df[i,])
  
} else if (grepl("3g", rownames(otutable_df)[i]) == T){
  otutable_df_D19_glass = rbind(otutable_df_D19_glass, otutable_df[i,])
  
} else if (grepl("4g", rownames(otutable_df)[i]) == T){
  otutable_df_D28_glass = rbind(otutable_df_D28_glass, otutable_df[i,])
  
} else if (grepl("5g", rownames(otutable_df)[i]) == T){
  otutable_df_D42_glass = rbind(otutable_df_D42_glass, otutable_df[i,])
  
} else if (grepl("6g", rownames(otutable_df)[i]) == T){
  otutable_df_D56_glass = rbind(otutable_df_D56_glass, otutable_df[i,])

} else if (grepl("1t", rownames(otutable_df)[i]) == T){
  otutable_df_D7_tile = rbind(otutable_df_D7_tile, otutable_df[i,]) 
  
} else if (grepl("2t", rownames(otutable_df)[i]) == T){
  otutable_df_D14_tile = rbind(otutable_df_D14_tile, otutable_df[i,])
  
} else if (grepl("3t", rownames(otutable_df)[i]) == T){
  otutable_df_D19_tile = rbind(otutable_df_D19_tile, otutable_df[i,])
  
} else if (grepl("4t", rownames(otutable_df)[i]) == T){
  otutable_df_D28_tile = rbind(otutable_df_D28_tile, otutable_df[i,])
  
} else if (grepl("5t", rownames(otutable_df)[i]) == T){
  otutable_df_D42_tile = rbind(otutable_df_D42_tile, otutable_df[i,])
  
} else if (grepl("6t", rownames(otutable_df)[i]) == T){
  otutable_df_D56_tile = rbind(otutable_df_D56_tile, otutable_df[i,])

} else if (grepl("1p", rownames(otutable_df)[i]) == T){
  otutable_df_D7_wood = rbind(otutable_df_D7_wood, otutable_df[i,]) 
  
} else if (grepl("2p", rownames(otutable_df)[i]) == T){
  otutable_df_D14_wood = rbind(otutable_df_D14_wood, otutable_df[i,])
  
} else if (grepl("3p", rownames(otutable_df)[i]) == T){
  otutable_df_D19_wood = rbind(otutable_df_D19_wood, otutable_df[i,])
  
} else if (grepl("4p", rownames(otutable_df)[i]) == T){
  otutable_df_D28_wood = rbind(otutable_df_D28_wood, otutable_df[i,])
  
} else if (grepl("5p", rownames(otutable_df)[i]) == T){
  otutable_df_D42_wood = rbind(otutable_df_D42_wood, otutable_df[i,])
  
} else if (grepl("6p", rownames(otutable_df)[i]) == T){
  otutable_df_D56_wood = rbind(otutable_df_D56_wood, otutable_df[i,])

} 
  i=i+1
}

#Check rownames to see if it worked properly
rownames(otutable_df)
rownames(otutable_df_D7_plastic)
rownames(otutable_df_D14_plastic)
rownames(otutable_df_D19_plastic)
rownames(otutable_df_D28_plastic)
rownames(otutable_df_D42_plastic)
rownames(otutable_df_D56_plastic)
rownames(otutable_df_D7_glass)
rownames(otutable_df_D14_glass)
rownames(otutable_df_D19_glass)
rownames(otutable_df_D28_glass)
rownames(otutable_df_D42_glass)
rownames(otutable_df_D56_glass)
rownames(otutable_df_D7_tile)
rownames(otutable_df_D14_tile)
rownames(otutable_df_D19_tile)
rownames(otutable_df_D28_tile)
rownames(otutable_df_D42_tile)
rownames(otutable_df_D56_tile)
rownames(otutable_df_D7_wood)
rownames(otutable_df_D14_wood)
rownames(otutable_df_D19_wood)
rownames(otutable_df_D28_wood)
rownames(otutable_df_D42_wood)
rownames(otutable_df_D56_wood)



#Specify order number with however many samples there are
substrate_list = c("plastic",
              "glass",
              "tile",
              "wood")
Time_list = c(7,
             14,
             19,
             28,
             42,
             56)
countordermax = list()
general_count = 1
#Run calculations
for (y in 1:4){
  for (i in 1:6) {
    Object = get(paste0("otutable_df_D", Time_list[i], "_", substrate_list[y]))
    
    temp = nrow(Object)
    countordermax[general_count] = temp
    general_count = general_count + 1
  }
 }

countordermax

general_count = 1
  
i=1
y=1

#Make progress bar and run ASV sharing calculations
pb = txtProgressBar(min = 0, max = length(countordermax), initial = 0, style = 3)
TIME <- Sys.time()
TIME
for (y in 1:4){
  for (i in 1:6) {
    Object = get(paste0("otutable_df_D", Time_list[i], "_", substrate_list[y]))
    zeta_decline.mc_calc = Zeta.decline.mc(data.spec =  Object, order = 1:countordermax[[general_count]])
    saveRDS(zeta_decline.mc_calc, paste0("otutable_df_18S_D", Time_list[i], "_", substrate_list[y]))

    Object2 = zeta_decline.mc_calc
      assign(paste0("otutable_df_D", Time_list[i], "_", substrate_list[y], "_18S_out"), Object2)
      
      setTxtProgressBar(pb, general_count)
      Sys.time() - TIME
      general_count = general_count + 1 
  }
  
}
Sys.time() - TIME
```
##Extract data
```{r}
#Make extraction list
Time_list = c(7,
              14,
              19,
              28,
              42,
              56)
substrate_list = c("plastic",
                   "glass",
                   "tile",
                   "wood")
Org_list = c("16S", "18S")

#Extarct data
Summary_Zeta = data.frame(matrix(ncol = 5, nrow = 0))
x = c("Day", "Substrate", "Shared", "Unique", "Sequencing")
colnames(Summary_Zeta) <- x
#pb = txtProgressBar(min = 0, max = length(Time_list), initial = 0)
i = 1

for (i in 1:length(Time_list)) {
  for (y in 1:length(substrate_list)) {
    for (x in 1:length(Org_list)) {
    
    assign("temp", get(paste0("otutable_df_D", Time_list[i], "_", substrate_list[y], "_", Org_list[x], "_out")))
    
    temp_df = data.frame("Day" = Time_list[i], 
                         "Substrate" = substrate_list[y],
                         "Shared" = tail(temp$zeta.val, n = 1), 
                         "Unique" = head(temp$zeta.val, n = 1), 
                         "Sequencing" = Org_list[x])
    
    Summary_Zeta = rbind(Summary_Zeta, temp_df)

#setTxtProgressBar(pb,i)
    }
  }
}
Summary_Zeta
Summary_Zeta$Sequencing = gsub("16S", "Prokaryotes", Summary_Zeta$Sequencing)
Summary_Zeta$Sequencing = gsub("18S", "Eukaryotes", Summary_Zeta$Sequencing)

Summary_Zeta$Substrate = gsub("plastic", "Plastic", Summary_Zeta$Substrate)
Summary_Zeta$Substrate = gsub("glass", "Glass", Summary_Zeta$Substrate)
Summary_Zeta$Substrate = gsub("tile", "Tile", Summary_Zeta$Substrate)
Summary_Zeta$Substrate = gsub("wood", "Wood", Summary_Zeta$Substrate)

Summary_Zeta
```
##Plot
```{r}


#Convert to factor and reorder for easier plotting, so plotting makes logical sense, and consistency
Summary_Zeta$Day = as.factor(as.character(Summary_Zeta$Day))
Summary_Zeta$Day = factor(Summary_Zeta$Day,
                         levels=c("7",
                                  "14",
                                  "19",
                                  "28",
                                  "42",
                                  "56"))
Summary_Zeta$Sequencing = factor(Summary_Zeta$Sequencing,
                         levels=c("Prokaryotes", 
                                  "Eukaryotes"))

Summary_Zeta$Substrate <- factor(Summary_Zeta$Substrate,
                         levels=c("Plastic", "Glass", "Tile", "Wood"))

#Make plot
Shared_plot_mesh = ggplot(data = Summary_Zeta, aes(x = Day, y = Shared, colour = Substrate, group = Substrate))+
  geom_line(lwd = 1)+
  scale_color_manual("Substrates", 
                     values = Substrate_colour_list)+
  scale_y_log10()+
  xlab("Time (days)")+
  ylab("Number of shared ASVs")+
 # ggtitle("Shared ASVs across samples increases with time")+
  facet_grid(. ~ Sequencing)+
  theme_bw()+
  My_Theme

#Show plot
Shared_plot_mesh


```
##Calculate means and statistical significance
```{r Set up}
#Subset to prokaryotes and eukaryotes specifically
Summary_Zeta_Prok = subset(Summary_Zeta, Sequencing == "Prokaryotes")
Summary_Zeta_Euk = subset(Summary_Zeta, Sequencing == "Eukaryotes")

#Add early vs late as new column
Summary_Zeta_Prok$StageofBiofilm = "test"
Summary_Zeta_Prok$StageofBiofilm[grep("7", Summary_Zeta_Prok$Day)] = "Early"
Summary_Zeta_Prok$StageofBiofilm[grep("14", Summary_Zeta_Prok$Day)] = "Early"
Summary_Zeta_Prok$StageofBiofilm[grep("19", Summary_Zeta_Prok$Day)] = "Late"
Summary_Zeta_Prok$StageofBiofilm[grep("28", Summary_Zeta_Prok$Day)] = "Late"
Summary_Zeta_Prok$StageofBiofilm[grep("42", Summary_Zeta_Prok$Day)] = "Late"
Summary_Zeta_Prok$StageofBiofilm[grep("56", Summary_Zeta_Prok$Day)] = "Late"
Summary_Zeta_Euk$StageofBiofilm = "tmp"
Summary_Zeta_Euk$StageofBiofilm[grep("7", Summary_Zeta_Euk$Day)] = "Early"
Summary_Zeta_Euk$StageofBiofilm[grep("14", Summary_Zeta_Euk$Day)] = "Late"
Summary_Zeta_Euk$StageofBiofilm[grep("19", Summary_Zeta_Euk$Day)] = "Late"
Summary_Zeta_Euk$StageofBiofilm[grep("28", Summary_Zeta_Euk$Day)] = "Late"
Summary_Zeta_Euk$StageofBiofilm[grep("42", Summary_Zeta_Euk$Day)] = "Late"
Summary_Zeta_Euk$StageofBiofilm[grep("56", Summary_Zeta_Euk$Day)] = "Late"

Summary_Zeta = rbind(Summary_Zeta_Prok, Summary_Zeta_Euk)
```


```{r Calculate means}
test_mean = Summary_Zeta_Prok %>% 
  group_by(Day) %>% 
  summarise(mean_val = mean(Shared), sd = sd(Shared))
print(test_mean)
#Mean increases across study
test_mean$mean_val[6]/test_mean$mean_val[1]

test_mean = Summary_Zeta_Euk %>% 
  group_by(Day) %>% 
  summarise(mean_val = mean(Shared), sd = sd(Shared))
print(test_mean)
#Mean increases across study
test_mean$mean_val[6]/test_mean$mean_val[1]

test_mean = Summary_Zeta %>% 
  group_by(Day) %>% 
  summarise(mean_val = mean(Shared), sd = sd(Shared))
print(test_mean)
```


```{r Identify statistical significance}
kruskal.test(Summary_Zeta$Shared, Summary_Zeta$Day, p.adjust.method = "bonferroni")
#Significant
wilcox.test(Summary_Zeta$Shared ~ Summary_Zeta$StageofBiofilm, p.adjust.method = "bonferroni")
#Significant

kruskal.test(Summary_Zeta_Prok$Shared, Summary_Zeta_Prok$Day, p.adjust.method = "bonferroni")
#Significant
kruskal.test(Summary_Zeta_Euk$Shared, Summary_Zeta_Euk$Day, p.adjust.method = "bonferroni")
#Non-significant

wilcox.test(Summary_Zeta_Prok$Shared ~ Summary_Zeta_Prok$StageofBiofilm, p.adjust.method = "bonferroni")
#Significant
wilcox.test(Summary_Zeta_Euk$Shared ~ Summary_Zeta_Euk$StageofBiofilm, p.adjust.method = "bonferroni")
#Significant

pairwise.wilcox.test(Summary_Zeta_Prok$Shared, Summary_Zeta_Prok$Day, p.adjust.method = "bonferroni")
#No significance
pairwise.wilcox.test(Summary_Zeta_Euk$Shared, Summary_Zeta_Euk$Day, p.adjust.method = "bonferroni")
#No significance

cor.test(as.numeric(as.character(Summary_Zeta_Prok$Shared)), as.numeric(as.character(Summary_Zeta_Prok$Day)), method = "s")
#Significant
cor.test(as.numeric(as.character(Summary_Zeta_Euk$Shared)), as.numeric(as.character(Summary_Zeta_Euk$Day)), method = "s")
#Non-significant

```


#Combine into one figure
```{r}

#Make
Figure6 = ggarrange(Shared_plot_substrate,
                   Shared_plot_mesh,
                common.legend = F,
                legend = "right",
                labels = c("A", "B"))
#Ensure it looks good
Figure6


#Save plot
pdf("~/Desktop/Biofilm_project/Analysis/Clean_Thesis_Code/Figures/Figure6_BiofilmASVSharing.pdf", width = 16, height = 4)
Figure6
dev.off()

```
