==> 8_Figure8_SignificantFamilyChanges.Rmd <==
---
title: "Figure 6 - Significant Family changes"
author: "Sven Tobias-Hunefeldt"
date: "30/03/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
#Set up
Import data and refine as done using the set up script

```{r load packages}

library(ggpubr)
library(ggplot2)
library(Rmisc)
library(vegan)
library(plyr)
library(dplyr)
library(phyloseq)
library(forcats)

```
```{r Family specific colours}

Family_colour_list <- c(#"Arthropoda" =  "black",
  #Ciliophora
                "Oligohymenophorea" = "red",
                "Phyllopharyngea" = "firebrick",
                "Hypotrichia" = "tomato",
                "Oligotrichia" = "indianred",
  #Cercozoa
                "Rhizaspididae" = "violet",
  #Euglenozoa 
                "Neobodonida" = "violetred4", 
                
  #Florideophycidae
                "Rhodymeniophycidae_fa" = "cornflowerblue",
                
        #"Incertae_Sedis_ph" = "",
        #"Nematoda" = "thistle3", 
  #Ochrophyta
                "Bacillariophyceae" = "green",
                "Fragilariales" = "forestgreen",
                "Ectocarpales_fa" = "darkolivegreen4",
                "Laminariales_fa" = "springgreen3",
  #Bacteroidetes
                "Saprospiraceae" = "violetred",
                "Flavobacteriaceae" = "magenta",
  #Proteobacteria
                "Rhodobacteraceae" = "navy",
                "Mitochondria" = "lightseagreen",
                "Colwelliaceae" = "deepskyblue",
                "Cellvibrionaceae" = "blue",
                "Halieaceae" = "turquoise",
                #"Epsilonbacteraeota" = "orange",
        #"Planctomycetes" = "beige",
        #"Proteobacteria" = "navy",
                
        #"Verrucomicrobia" = "magenta"
  "Rare_Taxa" = "black",
  "Rare Taxa (< 3%)" = "black"
                )

```


#Figure 6 A - significant prokaryotic family changes
```{r Set up}

#Restore a pristine copy from back up
Phyloseq_Biofilm_16S = Phyloseq_Biofilm_16S_v0

#Subset to only substrate associated biofilms
Phyloseq_Biofilm_16S_noA = subset_samples(Phyloseq_Biofilm_16S, Project == "Marine_Biofilm" & Sample_type == "Biofilm" & !Substrate == "Water"& !Substrate == "Mesh")

#Group them by their sample time
dge_EdgeR_obj_Time = phyloseq_to_edgeR(Phyloseq_Biofilm_16S_noA, group = "Sample_time")

# Perform binary test
et_EdgeR_Time = exactTest(dge_EdgeR_obj_Time)
# Adjust p value and extract values from test results
tt_EdgeR_Time = topTags(et_EdgeR_Time, n = nrow(dge_EdgeR_obj_Time$table), adjust.method = "BH", sort.by = "PValue")
res_EdgeR_Time = tt_EdgeR_Time@.Data[[1]]
sigtab_2fold_EdgeR_Time<- subset(res_EdgeR_Time, PValue < 0.05 & logFC >= 2 | PValue < 0.05 & logFC <= -2)


#Keep only FDR corrected <.1
sigtab_2fold_FDR_Time <- subset(sigtab_2fold_EdgeR_Time, FDR < 0.1)

keepTaxa_FDR_Time <- sigtab_2fold_EdgeR_Time$Genus #Extract the OTU table that was shown to be significant
Twofold_FDR_Time <- subset_taxa(Phyloseq_Biofilm_16S_noA, Genus %in% keepTaxa_FDR_Time) #Subset the taxa by the OTUs that were shown to change significantly
dat_2fold_FDR_Time <- tax_glom(Twofold_FDR_Time, taxrank = 'Family') %>%#Merge the species at the Family level
  transform_sample_counts(function(x) {x/sum(x)} ) %>% #Extract abundance data from phyloseq object
  psmelt() #Melt it into a dataframe

dat_2fold_FDR_Time <- dat_2fold_FDR_Time[order(dat_2fold_FDR_Time$Family),] #Order them at the Family level

dat_2fold_FDR_Time$Family <- as.character(dat_2fold_FDR_Time$Family)
  # group dataframe by Phylum, calculate relative abundance

medians_Time <- ddply(dat_2fold_FDR_Time, ~Family, function(x) c(median=mean(x$Abundance)))
  # find Phyla whose rel. abund. is less than 1%
remainder_Time <- medians_Time[medians_Time$median <= 0.03,]$Family
  # change their name to "Rare_Taxa"
dat_2fold_FDR_Time[dat_2fold_FDR_Time$Family %in% remainder_Time,]$Family <- 'Rare_Taxa'

#Sumamrise for plotting
Summary_Time_Family <- summarySE(dat_2fold_FDR_Time, measurevar="Abundance", groupvars=c("Family", "Mesh_status", "Sample_time", "StageofBiofilm", "Substrate"))
Summary_Time_Family

#write.csv(Summary_Time_Family, file = "Summary_Time_Family_16S.csv")

#Rearrange for easier plotting and dataframe browsing
Summary_Time_Family<-dplyr::arrange(Summary_Time_Family,Family, Abundance)
#Convert for easy renaming
Summary_Time_Family$Family = as.character(Summary_Time_Family$Family)
#Rename rare taxa with a more informative name
Summary_Time_Family$Family = gsub("Rare_Taxa", "Rare Taxa (< 3%)", Summary_Time_Family$Family)

#Remove excess family levels
Summary_Time_Family$Family <- factor(Summary_Time_Family$Family,
                         levels=(unique(Summary_Time_Family$Family)))

#Check for weird names and make Rare taxa name better.
levels(Summary_Time_Family$Family)
levels(Summary_Time_Family$Phylum)

```
```{r Plot significantly correlated families}

Summary_Time_Family$Mesh_status = gsub("Exposed", "Non-enclosed", Summary_Time_Family$Mesh_status)
#Reorder to make plotting easier
Summary_Time_Family$Sample_time = as.factor(as.character(Summary_Time_Family$Sample_time))
Summary_Time_Family$Sample_time = factor(Summary_Time_Family$Sample_time,
                         levels=c("7",
                                  "14",
                                  "19",
                                  "28",
                                  "42",
                                  "56"))

Summary_Time_Family$Substrate = factor(Summary_Time_Family$Substrate,
                         levels=c("Plastic",
                                  "Glass",
                                  "Tile",
                                  "Wood"))

Summary_plot_Time_16S = ggplot(Summary_Time_Family, 
                        aes(x=Sample_time, 
                            y=Abundance*100, 
                            colour= fct_reorder(Family, Abundance, .fun = mean, .desc = T),
                            group = Family))+ 
  geom_line(stat = "identity", lwd = 1) + 
  geom_errorbar(aes(ymin=(Abundance-se)*100, 
                    ymax=(Abundance+se)*100,
                    colour=Family), 
                position=pd
                )+
  theme_bw()+
  theme(legend.position = "bottom", 
        strip.background = element_blank())+
  scale_color_manual("Prokaryotes", values = Family_colour_list)+
  xlab("Time (days)")+
  ylab("Relative abundance (%)")+
  facet_grid(Substrate~Mesh_status)+
  theme_bw()+
  My_Theme
#Make sure it worked
Summary_plot_Time_16S


```
```{r Calculate means and some statistics}

test_mean = Summary_Time_Family %>% 
  group_by(Family) %>% 
  summarise(sum_val = sum(Abundance), mean_val = mean(Abundance), sd = sd(Abundance))
print(test_mean)

test_mean = Summary_Time_Family %>% 
  group_by(Family, Sample_time) %>% 
  summarise(mean_val = mean(Abundance), sd = sd(Abundance))
print(test_mean)

test_mean = Summary_Time_Family %>% 
  group_by(Family, Sample_time, Mesh_status) %>% 
  dplyr::summarise(mean_val = mean(Abundance), sd = sd(Abundance))
print(test_mean)

#Day 56 Flavo difference between enclosed and exposed
test_mean$mean_val[36] - test_mean$mean_val[35]

test_mean = Summary_Time_Family %>% 
  group_by(Family, Sample_time, Substrate) %>% 
  summarise(mean_val = mean(Abundance), sd = sd(Abundance))
print(test_mean, n = 48)

#Celvibrio pine abudnnace shifts
test_mean$mean_val[7] - test_mean$mean_val[3]

test_mean = Summary_Time_Family %>% 
  group_by(Family, Substrate) %>% 
  dplyr::summarise(mean_val = mean(Abundance), sd = sd(Abundance))
print(test_mean)

#Colwelliaceae variability based on substrates


test_mean = Summary_Time_Family %>% 
  group_by(Family, StageofBiofilm, Substrate) %>% 
  dplyr::summarise(mean_val = mean(Abundance), sd = sd(Abundance))
print(test_mean)

#Significant stage effect of Halieaceae
wilcox.test(Abundance ~ StageofBiofilm, data = Summary_Time_Family[c(142:188),])
#Significant

summary(aov(Abundance ~ StageofBiofilm * Substrate * Mesh_status, data = Summary_Time_Family[c(142:188),]))
#Significant sample time and substrate

summary(aov(Abundance ~ Sample_time * Substrate * Mesh_status, data = Summary_Time_Family[c(142:188),]))
#Significant sample time and substrate

pairwise.wilcox.test(Summary_Time_Family[c(142:188),]$Abundance, Summary_Time_Family[c(142:188),]$Substrate, p.adjust.method = "bonferroni")
#No significant differences

```
#Figure 6 B - significant eukaryotic family changes
```{r Set up}

#Restore a pristine copy from back up
Phyloseq_Biofilm_18S = Phyloseq_Biofilm_18S_v0

#Subset to only substrate associated biofilms
Phyloseq_Biofilm_18S_noA = subset_samples(Phyloseq_Biofilm_18S, Project == "Marine_Biofilm" & Sample_type == "Biofilm" & !Substrate == "Water"& !Substrate == "Mesh")

#Group them by their sample time
dge_EdgeR_obj_Time = phyloseq_to_edgeR(Phyloseq_Biofilm_18S_noA, group = "Sample_time")

# Perform binary test
et_EdgeR_Time = exactTest(dge_EdgeR_obj_Time)
# Adjust p value and extract values from test results
tt_EdgeR_Time = topTags(et_EdgeR_Time, n = nrow(dge_EdgeR_obj_Time$table), adjust.method = "BH", sort.by = "PValue")
res_EdgeR_Time = tt_EdgeR_Time@.Data[[1]]
sigtab_2fold_EdgeR_Time<- subset(res_EdgeR_Time, PValue < 0.05 & logFC >= 2 | PValue < 0.05 & logFC <= -2)


#Keep only FDR corrected <.1
sigtab_2fold_FDR_Time <- subset(sigtab_2fold_EdgeR_Time, FDR < 0.1)

keepTaxa_FDR_Time <- sigtab_2fold_EdgeR_Time$Genus #Extract the OTU table that was shown to be significant
Twofold_FDR_Time <- subset_taxa(Phyloseq_Biofilm_18S_noA, Genus %in% keepTaxa_FDR_Time) #Subset the taxa by the OTUs that were shown to change significantly
dat_2fold_FDR_Time <- tax_glom(Twofold_FDR_Time, taxrank = 'Family') %>%#Merge the species at the Family level
  transform_sample_counts(function(x) {x/sum(x)} ) %>% #Extract abundance data from phyloseq object
  psmelt() #Melt it into a dataframe

dat_2fold_FDR_Time <- dat_2fold_FDR_Time[order(dat_2fold_FDR_Time$Family),] #Order them at the Family level

dat_2fold_FDR_Time$Family <- as.character(dat_2fold_FDR_Time$Family)
  # group dataframe by Phylum, calculate relative abundance

#Calculate mean family abundance
medians_Time <- ddply(dat_2fold_FDR_Time, ~Family, function(x) c(median=mean(x$Abundance)))
  # find Phyla whose rel. abund. is less than 1%
remainder_Time <- medians_Time[medians_Time$median <= 0.03,]$Family
  # change their name to "Rare_Taxa"
dat_2fold_FDR_Time[dat_2fold_FDR_Time$Family %in% remainder_Time,]$Family <- 'Rare_Taxa'

#Summarise for plotting
Summary_Time_Family <- summarySE(dat_2fold_FDR_Time, measurevar="Abundance", groupvars=c("Family", "Mesh_status", "Sample_time", "Substrate"))
#Make sure it worked
Summary_Time_Family
#For records
#write.csv(Summary_Time_Family, file = "Summary_Time_Family_18S.csv")

#Rearrange for easier plotting and dataframe browsing
Summary_Time_Family<-dplyr::arrange(Summary_Time_Family,Family, Abundance)
#Convert for easy renaming
Summary_Time_Family$Family = as.character(Summary_Time_Family$Family)
#Rename rare taxa with a more informative name
Summary_Time_Family$Family = gsub("Rare_Taxa", "Rare Taxa (< 3%)", Summary_Time_Family$Family)

#Remove excess family levels
Summary_Time_Family$Family <- factor(Summary_Time_Family$Family,
                         levels=(unique(Summary_Time_Family$Family)))

#Check for weird names and make Rare taxa name better.
levels(Summary_Time_Family$Family)
levels(Summary_Time_Family$Phylum)

```
```{r Plot significantly correlated families}
Summary_Time_Family$Mesh_status = as.character(Summary_Time_Family$Mesh_status)
Summary_Time_Family$Mesh_status = gsub("Exposed", "Non-enclosed", Summary_Time_Family$Mesh_status)

#Reorder for consistent plotting
Summary_Time_Family$Sample_time = as.factor(as.character(Summary_Time_Family$Sample_time))
Summary_Time_Family$Sample_time = factor(Summary_Time_Family$Sample_time,
                         levels=c("7",
                                  "14",
                                  "19",
                                  "28",
                                  "42",
                                  "56"))
Summary_Time_Family$Substrate = factor(Summary_Time_Family$Substrate,
                         levels=c("Plastic",
                                  "Glass",
                                  "Tile",
                                  "Wood"))
#Generate plot design
Summary_plot_Time_18S = ggplot(Summary_Time_Family, 
                        aes(x=Sample_time, 
                            y=Abundance*100, 
                            colour = fct_reorder(Family, Abundance, .fun = mean, .desc = T),
                            group = Family))+ 
  geom_line(stat = "identity", lwd = 1) + 
  geom_errorbar(aes(ymin=(Abundance-se)*100, 
                    ymax=(Abundance+se)*100,
                    colour=Family), 
                position=pd
                )+
  theme_bw()+
  theme(legend.position = "bottom", 
        strip.background = element_blank())+
  scale_color_manual("Eukaryotes", values = Family_colour_list)+
  xlab("Time (days)")+
  ylab("Relative abundance (%)")+
  facet_grid(Substrate~Mesh_status)+
  theme_bw()+
  My_Theme
#Make sure it worked  
Summary_plot_Time_18S

```
```{r Calculate means}

test_mean = dat_2fold_FDR_Time %>% 
  group_by(Family, Sample_time, Mesh_status) %>% 
  dplyr::summarise(mean_val = mean(Abundance), sd = sd(Abundance))
print(test_mean)

test_mean = dat_2fold_FDR_Time %>% 
  group_by(Family, StageofBiofilm, Substrate) %>% 
  dplyr::summarise(mean_val = mean(Abundance), sd = sd(Abundance))
print(test_mean)


test_mean = dat_2fold_FDR_Time %>% 
  group_by(Family, StageofBiofilm, Mesh_status) %>% 
  dplyr::summarise(mean_val = mean(Abundance), sd = sd(Abundance))
print(test_mean)

```
#Combine for plot
```{r}

#Plot top
Figure8 = ggarrange(Summary_plot_Time_16S,
                Summary_plot_Time_18S,
                labels = c("A", "B"),
                ncol = 1,
                nrow = 2,
                common.legend = F,
                legend = "right",
                align = "hv")
#Make sure it worked
Figure8

#Save plot
pdf("~/Desktop/Biofilm_project/Analysis/Clean_Thesis_Code/Figures/Figure8_SignFamilyChanges.pdf", width = 16, height = 13)
Figure8
dev.off()

```
