==> 14_SupplementaryFigure5_BiofilmWaterASVSharing.Rmd <==
---
title: "Environmental and biofilm ASV shairng"
author: "Sven Tobias-Hunefeldt"
date: "26/03/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
#Set up
Import data and refine as done using the set up script

```{r load packages}

library(ggpubr)
library(ggplot2)
library(Rmisc)
library(vegan)
library(plyr)
library(dplyr)
library(phyloseq)
library(zetadiv)
library(glm2)


```
#Number of retained ASVs
##Set up and run ASV
```{r}

####Prokaryotes ####
#Remove mesh samples
Phyloseq_Biofilm_16S_water = subset_samples(Phyloseq_Biofilm_16S, !Substrate == "Mesh")

#Extract otu table and convert to presence absence scheme
otutable_df = as.data.frame(as.matrix(otu_table(Phyloseq_Biofilm_16S_water)))
otutable_df[otutable_df>0]<-1


#Make empty dataframes for each time point
otutable_df_D0 = data.frame()
otutable_df_D7 = data.frame()
otutable_df_D14 = data.frame()
otutable_df_D19 = data.frame()
otutable_df_D28 = data.frame()
otutable_df_D42 = data.frame()
otutable_df_D56 = data.frame()

#Fill dataframes
i = 1
#Split original dataframe into separate dataframes based on sample days
for (i in 1:nrow(otutable_df)) {
if (grepl("0", rownames(otutable_df)[i]) == T) { #Loop for 1 in sample name, and if true run associated code, it not true then move onto next else.
  otutable_df_D0 = rbind(otutable_df_D0, otutable_df[i,]) #Extract row and combine into day specific dataframe.
  
} else if (grepl("1", rownames(otutable_df)[i]) == T) {
  otutable_df_D7 = rbind(otutable_df_D7, otutable_df[i,])
  
}else if (grepl("2", rownames(otutable_df)[i]) == T) {
  otutable_df_D14 = rbind(otutable_df_D14, otutable_df[i,])
  
} else if (grepl("3", rownames(otutable_df)[i]) == T) {
  otutable_df_D19 = rbind(otutable_df_D19, otutable_df[i,])
  
} else if (grepl("4", rownames(otutable_df)[i]) == T) {
  otutable_df_D28 = rbind(otutable_df_D28, otutable_df[i,])
  
} else if (grepl("5", rownames(otutable_df)[i]) == T) {
  otutable_df_D42 = rbind(otutable_df_D42, otutable_df[i,])
  
} else if (grepl("6", rownames(otutable_df)[i]) == T) {
  otutable_df_D56 = rbind(otutable_df_D56, otutable_df[i,])

}
  i=i+1
}

#Check rownames to see if it worked properly
rownames(otutable_df)
rownames(otutable_df_D0)
rownames(otutable_df_D7)
rownames(otutable_df_D14)
rownames(otutable_df_D19)
rownames(otutable_df_D28)
rownames(otutable_df_D42)
rownames(otutable_df_D56)

#Specify order number with however many samples there are
Time_list = c(0,
             7,
             14,
             19,
             28,
             42,
             56)

countordermax = list()
general_count = 1
 for (i in 1:7) {
    Object = get(paste0("otutable_df_D", Time_list[i]))
    
    temp = nrow(Object)
    countordermax[general_count] = temp
    general_count = general_count + 1
 }

countordermax

general_count = 1
  
i=1
y=1

#Set up progress bar and run calculations
pb = txtProgressBar(min = 0, max = length(countordermax), initial = 0, style = 3)
TIME <- Sys.time()
TIME
i=1
y=1
general_count = 1
for (i in 1:7) {
    Object = get(paste0("otutable_df_D", Time_list[i]))
    zeta_decline.mc_calc = Zeta.decline.mc(data.spec =  Object, order = 1:countordermax[[general_count]])
    saveRDS(zeta_decline.mc_calc, paste0("otutable_df_16S_D", Time_list[i]))

    Object2 = zeta_decline.mc_calc
      assign(paste0("otutable_df_D", Time_list[i], "_16S_out"), Object2)
      
      
      setTxtProgressBar(pb, general_count)
      Sys.time() - TIME
      general_count = general_count + 1 
  
}
Sys.time() - TIME


####Eukaryotes ####
#Remove mesh samples
Phyloseq_Biofilm_18S_water = subset_samples(Phyloseq_Biofilm_18S, !Substrate == "Mesh")

#Extract otu table and convert to presence absence scheme
otutable_df = as.data.frame(as.matrix(otu_table(Phyloseq_Biofilm_18S_water)))
otutable_df[otutable_df>0]<-1


#Make empty dataframes for each time point
otutable_df_D0 = data.frame()
otutable_df_D7 = data.frame()
otutable_df_D14 = data.frame()
otutable_df_D19 = data.frame()
otutable_df_D28 = data.frame()
otutable_df_D42 = data.frame()
otutable_df_D56 = data.frame()

#Fill dataframes
i = 1
#Split original dataframe into separate dataframes based on sample days
for (i in 1:nrow(otutable_df)) {
if (grepl("0", rownames(otutable_df)[i]) == T) { #Loop for 1 in sample name, and if true run associated code, it not true then move onto next else.
  otutable_df_D0 = rbind(otutable_df_D0, otutable_df[i,]) #Extract row and combine into day specific dataframe.
  
} else if (grepl("1", rownames(otutable_df)[i]) == T) {
  otutable_df_D7 = rbind(otutable_df_D7, otutable_df[i,])
  
}else if (grepl("2", rownames(otutable_df)[i]) == T) {
  otutable_df_D14 = rbind(otutable_df_D14, otutable_df[i,])
  
} else if (grepl("3", rownames(otutable_df)[i]) == T) {
  otutable_df_D19 = rbind(otutable_df_D19, otutable_df[i,])
  
} else if (grepl("4", rownames(otutable_df)[i]) == T) {
  otutable_df_D28 = rbind(otutable_df_D28, otutable_df[i,])
  
} else if (grepl("5", rownames(otutable_df)[i]) == T) {
  otutable_df_D42 = rbind(otutable_df_D42, otutable_df[i,])
  
} else if (grepl("6", rownames(otutable_df)[i]) == T) {
  otutable_df_D56 = rbind(otutable_df_D56, otutable_df[i,])

}
  i=i+1
}

#Check rownames to see if it worked properly
rownames(otutable_df)
rownames(otutable_df_D0)
rownames(otutable_df_D7)
rownames(otutable_df_D14)
rownames(otutable_df_D19)
rownames(otutable_df_D28)
rownames(otutable_df_D42)
rownames(otutable_df_D56)

#Specify order number with however many samples there are
Time_list = c(0,
             7,
             14,
             19,
             28,
             42,
             56)

countordermax = list()
general_count = 1
 for (i in 1:7) {
    Object = get(paste0("otutable_df_D", Time_list[i]))
    
    temp = nrow(Object)
    countordermax[general_count] = temp
    general_count = general_count + 1
 }

countordermax

general_count = 1
  
i=1
y=1

pb = txtProgressBar(min = 0, max = length(countordermax), initial = 0, style = 3)
TIME <- Sys.time()
TIME
i=1
y=1
general_count = 1
for (i in 1:7) {
    Object = get(paste0("otutable_df_D", Time_list[i]))
    zeta_decline.mc_calc = Zeta.decline.mc(data.spec =  Object, order = 1:countordermax[[general_count]])
    saveRDS(zeta_decline.mc_calc, paste0("otutable_df_18S_D", Time_list[i]))

    Object2 = zeta_decline.mc_calc
      assign(paste0("otutable_df_D", Time_list[i], "_18S_out"), Object2)
      
      
      setTxtProgressBar(pb, general_count)
      Sys.time() - TIME
      general_count = general_count + 1 
  
}
Sys.time() - TIME

  
```
##Extract data
```{r}
#Make extraction lists
Time_list = c(0, 7,14,19,28,42,56)
Org_list = c("16S", "18S")

#Extarct data
Summary_Zeta_Shared = data.frame(matrix(ncol = 4, nrow = 0))
x = c("Day", "Shared", "Unique", "Sequencing")
colnames(Summary_Zeta_Shared) <- x
i = 1

for (i in 1:length(Time_list)) {
    for (x in 1:length(Org_list)) {
    
    assign("temp", get(paste0("otutable_df_D", Time_list[i], "_", Org_list[x], "_out")))
    
    temp_df = data.frame("Day" = Time_list[i], 
                         "Shared" = tail(temp$zeta.val, n = 1), 
                         "Unique" = head(temp$zeta.val, n = 1), 
                         "Sequencing" = Org_list[x])
    
    Summary_Zeta_Shared = rbind(Summary_Zeta_Shared, temp_df)

    }
  }

Summary_Zeta_Shared
Summary_Zeta_Shared$Sequencing = gsub("16S", "Prokaryotes", Summary_Zeta_Shared$Sequencing)
Summary_Zeta_Shared$Sequencing = gsub("18S", "Eukaryotes", Summary_Zeta_Shared$Sequencing)

```
##Plot
```{r}
Organism_list = c("Prokaryotes" = "red",
                  "Eukaryotes" = "navy")
Summary_Zeta_Shared$Day = as.factor(as.character(Summary_Zeta_Shared$Day))

Summary_Zeta_Shared$Day = factor(Summary_Zeta_Shared$Day,
                         levels=c("0", 
                                  "7",
                                  "14",
                                  "19",
                                  "28",
                                  "42",
                                  "56"))
Summary_Zeta_Shared_no0 = subset(Summary_Zeta_Shared, Day!=0)

Summary_Zeta_Shared_no0$Sequencing <- factor(Summary_Zeta_Shared_no0$Sequencing,
                         levels=c("Prokaryotes", "Eukaryotes"))

Shared_plot = ggplot(data = Summary_Zeta_Shared_no0, aes(x = Day, y = Shared, colour = Sequencing, group = Sequencing))+
  geom_line(lwd = 1)+
  scale_color_manual("Organisms", values = Organism_list)+
 # scale_y_log10()+
  xlab("Time (days)")+
  ylab("Number of shared ASVs")+
 # ggtitle("Shared ASVs across samples increases with time")+
  theme_bw()+
  My_Theme
Shared_plot

pdf("~/Desktop/Biofilm_project/Analysis/Clean_Thesis_Code/Figures/SupplementaryFigure5_Shared_ASVs_Water.pdf", height = 6, width = 8)
Shared_plot
dev.off()

```
