==> 15_SupplementaryFigure6_EcotoneAssignment.Rmd <==
---
title: "Ecotone assignment"
author: "Sven Tobias-Hunefeldt"
date: "30/03/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
#Set up
Import data and refine as done using the set up script
```{r Install ecotone assigner}


#install.packages("/Users/sventobias-hunefeldt/Desktop/Biofilm_project/Analysis/EcotoneFinder_0.1.0.tar.gz", repos = NULL, type = "source")
```
```{r load packages}

library(EcotoneFinder)
library(colorspace)
library(Rmisc)
library(reshape2)
library(phyloseq)

```
#Set up dataframes
##Prokaryotes
```{r Prokaryotes}

#Restore a pristine copy from back up
Phyloseq_Biofilm_16S = Phyloseq_Biofilm_16S_v0

#Look otu table and metadata, are they mergable?
rownames(Phyloseq_Biofilm_16S@sam_data)
rownames(Phyloseq_Biofilm_16S@otu_table)
#Merge them
OTU_Meta_df_16S = merge(Phyloseq_Biofilm_16S@sam_data, Phyloseq_Biofilm_16S@otu_table, by = 0)
#Reorder factors for downstream analysis
OTU_Meta_df_16S$Sample_time = factor(OTU_Meta_df_16S$Sample_time, 
                                    levels = c("0", "7", "14", "19", "28", "42", "56"))

#Subset merged dataframe
JustBiofilm_test = subset(OTU_Meta_df_16S, Substrate == "Plastic" | Substrate == "Glass" | Substrate == "Tile" | Substrate == "Wood")

#Assign informative rownames to dataframe
rownames(JustBiofilm_test) = JustBiofilm_test$Row.names
#Check dataframe to see what needs removing
JustBiofilm_test[c(1:5), c(1:20)]
#Remove excess information
JustBiofilm_test[,c(1:12,14,15)] = NULL #Only keep sample time
#Check dataframe to see if it worked
JustBiofilm_test[c(1:5), c(1:20)]
#Melt into long format
JustBiofilm_long = melt(JustBiofilm_test, id.vars = "Sample_time")
#Check for errors
JustBiofilm_long[c(1:239), c(1:3)]

#Identify type of variable for downstream anlysis
typeof(JustBiofilm_long$Sample_time)
typeof(JustBiofilm_long$variable)
typeof(JustBiofilm_long$value)
#Change data type for easier analysis
JustBiofilm_long$Sample_time = as.numeric(as.character(JustBiofilm_long$Sample_time))
JustBiofilm_long$variable = as.factor(as.character(JustBiofilm_long$variable))
JustBiofilm_long$value = as.numeric(as.character(JustBiofilm_long$value))
#Check to see if it worked
typeof(JustBiofilm_long$Sample_time)
typeof(JustBiofilm_long$variable)
typeof(JustBiofilm_long$value)

#Make summary for identify trends
JustBiofilm_summary = summarySE(JustBiofilm_long, measurevar = "value", groupvars = c("Sample_time", "variable"))
#Check for errors
JustBiofilm_summary[c(1:100), c(1:7)]
test = "ASV3"
filter(JustBiofilm_summary, JustBiofilm_summary$variable %in% test)

#Convert data back into a wide format
JustBiofilm_wide_16S = reshape2::dcast(JustBiofilm_summary, Sample_time ~ variable, value.var = "value")
#Check for error
JustBiofilm_wide_16S[c(1:7), c(1:100)]

#Change names for downstream analysis
colnames(JustBiofilm_wide_16S) = gsub("ASV", "Sp.", colnames(JustBiofilm_wide_16S))
colnames(JustBiofilm_wide_16S) = gsub("Sample_time", "Distance", colnames(JustBiofilm_wide_16S))
#Check for errors
JustBiofilm_wide_16S[c(1:7), c(1:100)]
#Redefine for easier analysis
JustBiofilm_wide_16S$Distance <- factor(JustBiofilm_wide_16S$Distance,
                         levels=(unique(JustBiofilm_wide_16S$Distance)))

```
```{r Eukaryotes}

#Restore a pristine copy from back up
Phyloseq_Biofilm_18S = Phyloseq_Biofilm_18S_v0

#Look otu table and metadata, are they mergable?
rownames(Phyloseq_Biofilm_18S@sam_data)
rownames(Phyloseq_Biofilm_18S@otu_table)
#Merge them
OTU_Meta_df_18S = merge(Phyloseq_Biofilm_18S@sam_data, Phyloseq_Biofilm_18S@otu_table, by = 0)
#Reorder factors for downstream analysis
OTU_Meta_df_18S$Sample_time = factor(OTU_Meta_df_18S$Sample_time, 
                                    levels = c("0", "7", "14", "19", "28", "42", "56"))

#Subset merged dataframe
JustBiofilm_test = subset(OTU_Meta_df_18S, Substrate == "Plastic" | Substrate == "Glass" | Substrate == "Tile" | Substrate == "Wood")

#Assign informative rownames to dataframe
rownames(JustBiofilm_test) = JustBiofilm_test$Row.names
#Check dataframe to see what needs removing
JustBiofilm_test[c(1:5), c(1:20)]
#Remove excess information
JustBiofilm_test[,c(1:12,14,15)] = NULL #Only keep sample time
#Check dataframe to see if it worked
JustBiofilm_test[c(1:5), c(1:20)]
#Melt into long format
JustBiofilm_long = melt(JustBiofilm_test, id.vars = "Sample_time")
#Check for errors
JustBiofilm_long[c(1:239), c(1:3)]

#Identify type of variable for downstream anlysis
typeof(JustBiofilm_long$Sample_time)
typeof(JustBiofilm_long$variable)
typeof(JustBiofilm_long$value)
#Change data type for easier analysis
JustBiofilm_long$Sample_time = as.numeric(as.character(JustBiofilm_long$Sample_time))
JustBiofilm_long$variable = as.factor(as.character(JustBiofilm_long$variable))
JustBiofilm_long$value = as.numeric(as.character(JustBiofilm_long$value))
#Check to see if it worked
typeof(JustBiofilm_long$Sample_time)
typeof(JustBiofilm_long$variable)
typeof(JustBiofilm_long$value)

#Make summary for identify trends
JustBiofilm_summary = summarySE(JustBiofilm_long, measurevar = "value", groupvars = c("Sample_time", "variable"))
#Check for errors
JustBiofilm_summary[c(1:100), c(1:7)]
test = "ASV3"
filter(JustBiofilm_summary, JustBiofilm_summary$variable %in% test)

#Convert data back into a wide format
JustBiofilm_wide_18S = reshape2::dcast(JustBiofilm_summary, Sample_time ~ variable, value.var = "value")
#Check for error
JustBiofilm_wide_18S[c(1:7), c(1:100)]

#Change names for downstream analysis
colnames(JustBiofilm_wide_18S) = gsub("ASV", "Sp.", colnames(JustBiofilm_wide_18S))
colnames(JustBiofilm_wide_18S) = gsub("Sample_time", "Distance", colnames(JustBiofilm_wide_18S))
#Check for errors
JustBiofilm_wide_18S[c(1:7), c(1:100)]
#Redefine for easier analysis
JustBiofilm_wide_18S$Distance <- factor(JustBiofilm_wide_18S$Distance,
                         levels=(unique(JustBiofilm_wide_18S$Distance)))

```

#Run analyses
##Prokaryotes
```{r}
#### 2 cluster Analyses####
SyntheticEcoFinder_2_16S <- EcotoneFinder(data = JustBiofilm_wide_16S,
                                     dist = "Distance",
                                     method = "vegclust",
                                     groups = 2, 
                                     standardize = "hellinger",
                                     diversity = "all")
#Extract useful info for plotting
Forplot_2_16S = SyntheticEcoFinder_2_16S$vegclust$memb
#Generate plot
Plot_2_16S = ggplot(Forplot_2_16S, 
                    aes(x = as.factor(c(7, 14, 19, 28, 42, 56))))+
  geom_line(aes(y = M1, 
                colour = "red", 
                group = "M1"), 
            lwd = 1)+
  geom_line(aes( y = M2, 
                 colour = "magenta", 
                 group = "M2"),
            lwd = 1)+
  xlab("Time (days)")+
  ylab("Ecological gradient based \non Fuzzy Cluster Means")+
  theme_bw()+
  My_Theme+
  theme(axis.text.y = element_blank(),
        axis.title.x = element_blank(),
        legend.position = "none")
#Check plot
Plot_2_16S


#### 3 cluster Analyses####
SyntheticEcoFinder_3_16S <- EcotoneFinder(data = JustBiofilm_wide_16S,
                                     dist = "Distance",
                                     method = "vegclust",
                                     groups = 3, 
                                     standardize = "hellinger",
                                     diversity = "all")
#Extract useful info for plotting
Forplot_3_16S = SyntheticEcoFinder_3_16S$vegclust$memb
#Generate plot
Plot_3_16S = ggplot(Forplot_3_16S, 
                    aes(x = as.factor(c(7, 14, 19, 28, 42, 56))))+
  geom_line(aes(y = M1, 
                colour = "red", 
                group = "M1"),
            lwd = 1)+
  geom_line(aes( y = M2, 
                 colour = "green", 
                 group = "M2"),
            lwd = 1)+
  geom_line(aes( y = M3, 
                 colour = "magenta", 
                 group = "M3"),
            lwd = 1)+
  xlab("Time (days)")+
  ylab("Ecological gradient based \non Fuzzy Cluster Means")+
  theme_bw()+
  My_Theme+
  theme(axis.text.y = element_blank(),
        axis.title.x = element_blank(),
        legend.position = "none")
#Check plot
Plot_3_16S


#### 4 cluster Analyses####
SyntheticEcoFinder_4_16S <- EcotoneFinder(data = JustBiofilm_wide_16S,
                                     dist = "Distance",
                                     method = "vegclust",
                                     groups = 4, 
                                     standardize = "hellinger",
                                     diversity = "all")
#Extract useful info for plotting
Forplot_4_16S = SyntheticEcoFinder_4_16S$vegclust$memb
#Generate plot
Plot_4_16S = ggplot(Forplot_4_16S, 
                    aes(x = as.factor(c(7, 14, 19, 28, 42, 56))))+
  geom_line(aes(y = M1, 
                colour = "red", 
                group = "M1"),
            lwd = 1)+
  geom_line(aes( y = M2, 
                 colour = "green", 
                 group = "M2"),
            lwd = 1)+
  geom_line(aes( y = M3, 
                 colour = "magenta", 
                 group = "M3"),
            lwd = 1)+
    geom_line(aes( y = M4, 
                 colour = "blue", 
                 group = "M4"),
            lwd = 1)+
  xlab("Time (days)")+
  ylab("Ecological gradient based \non Fuzzy Cluster Means")+
  theme_bw()+
  My_Theme+
  theme(axis.text.y = element_blank(),
        legend.position = "none")
#Check plot
Plot_4_16S

```
##Eukaryotes
```{r}
#### 2 cluster Analyses####
SyntheticEcoFinder_2_18S <- EcotoneFinder(data = JustBiofilm_wide_18S,
                                     dist = "Distance",
                                     method = "vegclust",
                                     groups = 2, 
                                     standardize = "hellinger",
                                     diversity = "all")
#Extract useful info for plotting
Forplot_2_18S = SyntheticEcoFinder_2_18S$vegclust$memb
#Generate plot
Plot_2_18S = ggplot(Forplot_2_18S, 
                    aes(x = as.factor(c(7, 14, 19, 28, 42, 56))))+
  geom_line(aes(y = M1, 
                colour = "red", 
                group = "M1"),
            lwd = 1)+
  geom_line(aes( y = M2, 
                 colour = "magenta", 
                 group = "M2"),
            lwd = 1)+
  xlab("Time (days)")+
  ylab("Ecological gradient based on Fuzzy Cluster Means")+
  theme_bw()+
  My_Theme+
  theme(axis.text.y = element_blank(),
        axis.title.y = element_blank(),
        axis.title.x = element_blank(),
        legend.position = "none")
#Check plot
Plot_2_18S


#### 3 cluster Analyses####
SyntheticEcoFinder_3_18S <- EcotoneFinder(data = JustBiofilm_wide_18S,
                                     dist = "Distance",
                                     method = "vegclust",
                                     groups = 3, 
                                     standardize = "hellinger",
                                     diversity = "all")
#Extract useful info for plotting
Forplot_3_18S = SyntheticEcoFinder_3_18S$vegclust$memb
#Generate plot
Plot_3_18S = ggplot(Forplot_3_18S, 
                    aes(x = as.factor(c(7, 14, 19, 28, 42, 56))))+
  geom_line(aes(y = M1, 
                colour = "red", 
                group = "M1"),
            lwd = 1)+
  geom_line(aes( y = M2, 
                 colour = "green", 
                 group = "M2"),
            lwd = 1)+
  geom_line(aes( y = M3, 
                 colour = "magenta", 
                 group = "M3"),
            lwd = 1)+
  xlab("Time (days)")+
  ylab("Ecological gradient based on Fuzzy Cluster Means")+
  theme_bw()+
  My_Theme+
  theme(axis.text.y = element_blank(),
        axis.title.y = element_blank(),
        axis.title.x = element_blank(),
        legend.position = "none")
#Check plot
Plot_3_18S


#### 4 cluster Analyses####
SyntheticEcoFinder_4_18S <- EcotoneFinder(data = JustBiofilm_wide_18S,
                                     dist = "Distance",
                                     method = "vegclust",
                                     groups = 4, 
                                     standardize = "hellinger",
                                     diversity = "all")
#Extract useful info for plotting
Forplot_4_18S = as.data.frame(SyntheticEcoFinder_4_18S$vegclust$memb)
#Generate plot
Plot_4_18S = ggplot(Forplot_4_18S, 
                    aes(x = as.factor(c(7, 14, 19, 28, 42, 56))))+
  geom_line(aes(y = M1, 
                colour = "red", 
                group = "M1"),
            lwd = 1)+
  geom_line(aes( y = M2, 
                 colour = "green", 
                 group = "M2"),
            lwd = 1)+
  geom_line(aes( y = M3, 
                 colour = "magenta", 
                 group = "M3"),
            lwd = 1)+
    geom_line(aes( y = M4, 
                 colour = "blue", 
                 group = "M4"),
            lwd = 1)+
  xlab("Time (days)")+
  ylab("Ecological gradient based on Fuzzy Cluster Means")+
  theme_bw()+
  My_Theme+
  theme(axis.text.y = element_blank(),
        axis.title.y = element_blank(),
        legend.position = "none")
#Check plot
Plot_4_18S

```
#Make figure using ggarrange

```{r}

#Make figure
SuppFigure6 = ggarrange(Plot_2_16S, Plot_2_18S,
          Plot_3_16S, Plot_3_18S,
          Plot_4_16S, Plot_4_18S,
          ncol = 2,
          nrow = 3)
#Check for error
SuppFigure6

#Save figure
pdf("~/Desktop/Biofilm_project/Analysis/Clean_Thesis_Code/Figures/SupplementaryFigure6_EcotoneAssignment.pdf", height = 10, width = 10)
SuppFigure6
dev.off()

```
